Conflicts:
	Projects/MacOSFramework/Framework.xcodeproj/project.pbxproj

Former-commit-id: a1bba4cf9e3f889c659abdc075f26cdddcbf0b0e

+++ Projects/UnitTests/Classes/JPEGTest.cpp
@@ -132,7 +132,7 @@ void JPEGTest::ReloadSprites()
     Image* imgJpeg = imageSet[0];
     Texture* tex = Texture::CreateFromData(imgJpeg->GetPixelFormat(), imgJpeg->data, imgJpeg->width, imgJpeg->height, false);
     SafeRelease(imgJpeg);
-    jpegSprite = Sprite::CreateFromTexture(tex, 0, 0, tex->width, tex->width);
+    jpegSprite = Sprite::CreateFromTexture(tex, 0, 0, static_cast<float32>(tex->width), static_cast<float32>(tex->width));
     SafeRelease(tex);
 }
